added: integration tests for _search data streams
