--- integration-tests/src/test/resources/indices_aliases_test/readonlyrest.yml
@@ -1,38 +1,0 @@
-readonlyrest:
-  ssl:
-    enable: true
-    keystore_file: "keystore.jks"
-    keystore_pass: readonlyrest
-    key_pass: readonlyrest
-
-  prompt_for_basic_auth: false
-
-  access_control_rules:
-
-  - name: "::INITIALIZER::"
-    auth_key: admin:container
-
-  - name: "::GETTER::"
-    auth_key: "unrestricted:dev"
-
-  - name: "::RESTRICTED TO ALIAS::"
-    indices: ["public*"]
-    auth_key: "restricted:dev"
-
-  - name: "::PERFMON - RESTRICTED TO ALIAS::"
-    indices: ["perfmon*"]
-    auth_key: "perfmon:dev"
-
-  - name: "::PERFMON - RESTRICTED TO INDEX::"
-    indices: ["bla*"]
-    auth_key: "perfmon:dev"
-
-  - name: myanmar_vietnam
-    kibana_access: ro
-    auth_key: "VIET_MYAN:dev"
-    indices: ["*myanmar*", "*vietnam*"]
-
-  - name: "CONTAINER ADMIN"
-    verbosity: error
-    type: allow
-    auth_key: admin:container
