+++ integration-tests/src/test/scala/tech/beshu/ror/integration/proxy/SearchApiProxyTests.scala
@@ -14,9 +14,9 @@
  *    You should have received a copy of the GNU General Public License
  *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
  */
-package tech.beshu.ror.integration.plugin
+package tech.beshu.ror.integration.proxy
 
-import tech.beshu.ror.integration.suites.SearchWithAliasesSuite
-import tech.beshu.ror.integration.utils.SingletonPluginTestSupport
+import tech.beshu.ror.integration.suites.SearchApiSuite
+import tech.beshu.ror.integration.utils.XpackEsClusterProxyTestSupport
 
-class SearchWithAliasesPluginTests extends SearchWithAliasesSuite with SingletonPluginTestSupport
+class SearchApiProxyTests extends SearchApiSuite with XpackEsClusterProxyTestSupport
